backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic
python-dotenv
psycopg2-binary
